app/requirements.txt
flask
db-sqlite3
pycryptodome
google-api-python-client
python-dotenv